backend/requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Limiter==3.5.0
Flask-CORS==4.0.0
python-dotenv==1.0.0
bcrypt==4.1.2
WTForms==3.1.1
email-validator==2.1.0
scikit-learn==1.3.2
pandas==2.1.4